container: use /src as workdir
When the directory containing elbe is a child of /usr, elbe will not use
the elbe version from the disk but an official version online (it's the
goal of the `is_devel` option i version.py).
This commit change the default directory of the container from /usr/src
to /src.

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>

Signed-off-by: Thomas Bonnefille <thomas.bonnefille@bootlin.com>